breakout structs from single file
