+++ pkg/codegen/opsecdef.go
@@ -1,0 +1,6 @@
+package codegen
+
+type SecurityDefinition struct {
+	ProviderName string
+	Scopes       []string
+}
